Refactoring of main.py
